+++ Riot/SupportingFiles/Riot-Bridging-Header.h
@@ -27,3 +27,5 @@
 #import "JitsiWidgetData.h"
 #import "InviteRecentTableViewCell.h"
 #import "AuthFallBackViewController.h"
+#import "CallViewController.h"
+#import "MatrixContactsDataSource.h"